backend/requirements.txt
Django>=5.1,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.6,<5
djangorestframework-simplejwt>=5.3,<6
django-filter>=24.3,<26
